delete LoopCommand CostTime fix and Sub-Cmd timeline fix

+++ ReOsuStoryBoardPlayer/Base/CommandParameters.cs
@@ -142,7 +142,7 @@ namespace ReOsuStoryBoardPlayer
 
         public LoopCommandParamesters() { }
 
-        public uint CostTime;
+        public float CostTime;
     }
 
     public class ParameterCommandParamester : CommandParameters
